First attempt to handle PGC Green's functions
PGC GF are in a different format to Dreger's examples,
so modify code to handle these. This is an interim
step as the PGC can provide us with GF for all depths,
until we get on-demand creation of our own GFs within
Antelope.

This is also the first attempt to pull trace objects
from the database (wfdisc) rather than from hard-coded
files.

Right now there is a problem of getting enough values in
the event matrix that match the values in the GF. Until
those values match, we won't be able to generate the
moment tensor.

Also new are comments and notes that outline the problems,
including the author of the notes. Please retain this
format, adding your initials to the header comments