Finish composite image processing using PIL
Add a new item to the parameter file
to define path to true-type font (a 'feature'
of the Python Imaging Library that is
required for writing text) and finish the
building of the composite image (synthetics
plus focal mechanism). Move the db table update
to the composite image creation method.