Add clipping limits and smarter time series select
Modify the pf with the addition of a clipping
array of max and min counts. This is 80% of 2**23
as per personal comm with Frank Vernon.

Make the time series excerpt for each station
smarter, using the arrival.time of the P-wave
to determine the start and end times of the per
station trace object extracted and used in the
calculation.

Also, switch back in the pre-converted PGC
Green's function (in the time domain) for
testing.